added proposal.go to outsourece methods from types
