+++ x/cardchain/types/types.go
@@ -2,37 +2,10 @@ package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
 // Here can constructors for types be placed, sice the files containing those
 // types shall not be edited
-
-// Module name for minting coins
-const CoinsIssuerName = "cardchain"
-
-const (
-	// ProposalTypeChange defines the type for a ParameterChangeProposal
-	ProposalTypeCopyright = "Copyright"
-)
-
-func (c *CopyrightProposal) ProposalRoute() string { return RouterKey }
-
-func (c *CopyrightProposal) ProposalType() string { return ProposalTypeCopyright }
-
-func (c *CopyrightProposal) ValidateBasic() error {
-	err := govtypes.ValidateAbstract(c)
-	if err != nil {
-		return err
-	}
-	// TODO More validation
-	return nil
-}
-
-func init() {
-	govtypes.RegisterProposalType(ProposalTypeCopyright)
-	// govtypes.RegisterProposalTypeCodec(&CopyrightProposal{}, "cosmos-sdk/CopyrightProposal")
-}
 
 func NewUser() User {
 	return User{
